src/main/java/org/jtwig/model/expression/operation/calculators/binary/EquivalentOperationCalculator.java
package org.jtwig.model.expression.operation.calculators.binary;

import org.jtwig.context.RenderContext;
import org.jtwig.model.expression.Expression;
import org.jtwig.model.position.Position;
import org.jtwig.util.JtwigValue;

public class EquivalentOperationCalculator implements BinaryOperationCalculator {
    @Override
    public JtwigValue calculate(RenderContext context, Position position, Expression leftOperand, Expression rightOperand) {
        return new JtwigValue(leftOperand.calculate(context).equals(rightOperand.calculate(context)));
    }
}
